+++ src/main/java/com/TeachingManager/TeachingManager/controller/API/Schedule/ApiScheduleController.java
@@ -8,7 +8,9 @@ import com.TeachingManager.TeachingManager.domain.Schedule;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -22,48 +24,58 @@ public class ApiScheduleController {
     /*  일정 추가 api  */
     @PostMapping("/api/Schedule")
     public ResponseEntity<Schedule> Create_Schedule(@AuthenticationPrincipal CustomUser user, @RequestBody AddScheduleRequest request) {
-        Schedule sc = scheduleService.create_schedule(request, user.getPk());
-
-        if (sc != null){
-            return ResponseEntity.status(HttpStatus.CREATED).body(sc);
+        if(user != null && user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_PRESIDENT"))) {
+                return ResponseEntity.status(HttpStatus.CREATED).body(scheduleService.create_schedule(request, user));
         }
-        else{
-            // 존재하지 않는 이메일로 들어왔을 경우.
-            return (ResponseEntity<Schedule>) ResponseEntity.status(HttpStatus.BAD_REQUEST);
-        }
+        return ResponseEntity.badRequest().build();
     }
 
     // 자기 자신학원, 또는 소속되어있는 학원의 특정 달 정보를 가져온다. (이번달)
     @GetMapping("/api/Schedule")
     public ResponseEntity<MonthScheduleResponse> Schedule(@AuthenticationPrincipal CustomUser user, @RequestBody MonthScheduleRequest request){
-        return ResponseEntity.status(HttpStatus.CREATED).body(scheduleService.searchAll_scheduleByDate(user, request.getDate_info()));
+        if (user != null) {
+            return ResponseEntity.status(HttpStatus.CREATED).body(scheduleService.searchAll_scheduleByDate(user, request.getDate_info()));
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     /*     일정 디테일    */
     @GetMapping("/api/Schedule/{pk}")
     public ResponseEntity<ScheduleResponse> Detail_Schedule(@PathVariable("pk") Long pk,@AuthenticationPrincipal CustomUser user ) {
-        ScheduleResponse response = new ScheduleResponse(scheduleService.search_schedule(user, pk));
-        return ResponseEntity.ok()
-                .body(response);
+        if (user != null) {
+            return ResponseEntity.ok()
+                    .body(new ScheduleResponse(scheduleService.search_schedule(user, pk)));
+        }
+        return ResponseEntity.badRequest().build();
 
     }
 
     /*  일정 삭제 api  */
     @PostMapping("/api/delete/Schedule/{pk}")
     public ResponseEntity<Void> Delete_Schedule(@PathVariable("pk") Long pk, @AuthenticationPrincipal CustomUser user) {
-        scheduleService.delete_schedule(user, pk);
-        return ResponseEntity.ok().build();
+        // 학원 권한을 가진 쪽에서 요청인지
+        if(user != null && user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_PRESIDENT"))) {
+            scheduleService.delete_schedule(user, pk);
+            return ResponseEntity.ok().build();
+        }
+        // 아니라면 거절
+        return ResponseEntity.badRequest().build();
     }
 
     /*  일정 수정 api  */
     @PutMapping("/api/Schedule/{pk}")
     public ResponseEntity<ScheduleResponse> Revise_Schedule(@PathVariable("pk") Long pk, @RequestBody UpdateScheduleRequest request, @AuthenticationPrincipal CustomUser user) {
-        Schedule sc = scheduleService.update_schedule(user, pk, request);
-        return ResponseEntity.ok()
-                .body(new ScheduleResponse(sc));
+        // 학원 권한을 가진 쪽에서 요청이 들어왔는가
+        if(user != null && user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_PRESIDENT"))) {
+            Schedule sc = scheduleService.update_schedule(user, pk, request);
+            return ResponseEntity.ok()
+                    .body(new ScheduleResponse(sc));
+        }
+        
+        // 학원에서 들어오 요청이 아니라면.
+        return ResponseEntity.badRequest().build();
 
     }
-
-
+    
 }
